submited all the code from the computer for this app for the first time
this app is about people inputting their preference(color, passion, game, space,...) in the app so that their family and friend can discover and best gift them on birthday and similar event. afterward i can create an app or platform where people can additionally organize this event by agreeing on what gift to buy and the app can suggest it in the app which can be one way of revenue.